+++ sdk/metric/metricdata/metricdatatest/comparisons.go
@@ -698,7 +698,7 @@ func hasAttributesExemplars[T int64 | float64](
 			continue
 		}
 		if val != attr.Value {
-			reasons = append(reasons, notEqualStr(string(attr.Key), attr.Value.Emit(), val.Emit()))
+			reasons = append(reasons, notEqualStr(string(attr.Key), attr.Value.String(), val.String()))
 		}
 	}
 	return reasons
@@ -715,7 +715,7 @@ func hasAttributesDataPoints[T int64 | float64](
 			continue
 		}
 		if val != attr.Value {
-			reasons = append(reasons, notEqualStr(string(attr.Key), attr.Value.Emit(), val.Emit()))
+			reasons = append(reasons, notEqualStr(string(attr.Key), attr.Value.String(), val.String()))
 		}
 	}
 	return reasons
@@ -754,7 +754,7 @@ func hasAttributesHistogramDataPoints[T int64 | float64](
 			continue
 		}
 		if val != attr.Value {
-			reasons = append(reasons, notEqualStr(string(attr.Key), attr.Value.Emit(), val.Emit()))
+			reasons = append(reasons, notEqualStr(string(attr.Key), attr.Value.String(), val.String()))
 		}
 	}
 	return reasons
@@ -785,7 +785,7 @@ func hasAttributesExponentialHistogramDataPoints[T int64 | float64](
 			continue
 		}
 		if val != attr.Value {
-			reasons = append(reasons, notEqualStr(string(attr.Key), attr.Value.Emit(), val.Emit()))
+			reasons = append(reasons, notEqualStr(string(attr.Key), attr.Value.String(), val.String()))
 		}
 	}
 	return reasons
@@ -881,7 +881,7 @@ func hasAttributesSummaryDataPoint(dp metricdata.SummaryDataPoint, attrs ...attr
 			continue
 		}
 		if val != attr.Value {
-			reasons = append(reasons, notEqualStr(string(attr.Key), attr.Value.Emit(), val.Emit()))
+			reasons = append(reasons, notEqualStr(string(attr.Key), attr.Value.String(), val.String()))
 		}
 	}
 	return reasons